refactor: remove 2 layers constants
remove llmd_constants wrappers, use utilities.constants directly
